+++ PKHeX.Core/PKM/PK7.cs
@@ -103,12 +103,12 @@ namespace PKHeX.Core
         public override int EV_SPE { get => Data[0x21]; set => Data[0x21] = (byte)value; }
         public override int EV_SPA { get => Data[0x22]; set => Data[0x22] = (byte)value; }
         public override int EV_SPD { get => Data[0x23]; set => Data[0x23] = (byte)value; }
-        public int CNT_Cool { get => Data[0x24]; set => Data[0x24] = (byte)value; }
-        public int CNT_Beauty { get => Data[0x25]; set => Data[0x25] = (byte)value; }
-        public int CNT_Cute { get => Data[0x26]; set => Data[0x26] = (byte)value; }
-        public int CNT_Smart { get => Data[0x27]; set => Data[0x27] = (byte)value; }
-        public int CNT_Tough { get => Data[0x28]; set => Data[0x28] = (byte)value; }
-        public int CNT_Sheen { get => Data[0x29]; set => Data[0x29] = (byte)value; }
+        public byte CNT_Cool   { get => Data[0x24]; set => Data[0x24] = value; }
+        public byte CNT_Beauty { get => Data[0x25]; set => Data[0x25] = value; }
+        public byte CNT_Cute   { get => Data[0x26]; set => Data[0x26] = value; }
+        public byte CNT_Smart  { get => Data[0x27]; set => Data[0x27] = value; }
+        public byte CNT_Tough  { get => Data[0x28]; set => Data[0x28] = value; }
+        public byte CNT_Sheen  { get => Data[0x29]; set => Data[0x29] = value; }
         public byte ResortEventStatus { get => Data[0x2A]; set => Data[0x2A] = value; }
         private byte PKRS { get => Data[0x2B]; set => Data[0x2B] = value; }
         public override int PKRS_Days { get => PKRS & 0xF; set => PKRS = (byte)((PKRS & ~0xF) | value); }
